+++ crates/extensions_ui/src/components/extension_card.rs
@@ -4,21 +4,14 @@ use ui::prelude::*;
 
 #[derive(IntoElement)]
 pub struct ExtensionCard {
-    overridden_by_dev_extension: bool,
     children: SmallVec<[AnyElement; 2]>,
 }
 
 impl ExtensionCard {
     pub fn new() -> Self {
         Self {
-            overridden_by_dev_extension: false,
             children: SmallVec::new(),
         }
-    }
-
-    pub fn overridden_by_dev_extension(mut self, overridden: bool) -> Self {
-        self.overridden_by_dev_extension = overridden;
-        self
     }
 }
 
@@ -41,21 +34,7 @@ impl RenderOnce for ExtensionCard {
                 .border_1()
                 .border_color(cx.theme().colors().border_variant)
                 .rounded_md()
-                .children(self.children)
-                .when(self.overridden_by_dev_extension, |card| {
-                    card.child(
-                        h_flex()
-                            .absolute()
-                            .top_0()
-                            .left_0()
-                            .block_mouse_except_scroll()
-                            .cursor_default()
-                            .size_full()
-                            .justify_center()
-                            .bg(cx.theme().colors().elevated_surface_background.alpha(0.8))
-                            .child(Label::new("Overridden by dev extension.")),
-                    )
-                }),
+                .children(self.children),
         )
     }
 }